update game instructions and project comments
