+++ game/src/myCheckers/App.java
@@ -11,7 +11,6 @@ public class App extends Application
     public static void main(String[] args)
     {
         launch(args);
-
     }
 
     @Override
